requirements.txt
PyMuPDF==1.23.26
pandas==2.2.1
Pillow==10.2.0
fpdf==1.7.2
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
PyQt6==6.8.0
PyQt6-Qt6==6.8.1
PyQt6_sip==13.9.1
